Safari-609.1.11.1/LayoutTests/accessibility/mac/file-upload-button-subrole-expected.txt


This tests that a file upload button will have a specific subrole and a title that matches the files selected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS uploadButton.subrole is 'AXSubrole: AXFileUploadButton'
PASS uploadButton.roleDescription is 'AXRoleDescription: file upload button'
PASS uploadButton.title is 'AXTitle: no file selected'
PASS successfullyParsed is true

TEST COMPLETE
